refactor(server): rename LeaderBoardRoute.js → leaderBoardRoute.js and TodoRoute.js → todoRoute.js; update router const names
