+++ server/routes/leaderBoardRoute.js
@@ -1,0 +1,8 @@
+import express from "express";
+import { fetchLeaderboard } from "../controllers/leaderboardController.js";
+
+const leaderBoardRoute = express.Router();
+
+leaderBoardRoute.get("/leaderboard", fetchLeaderboard);
+
+export default leaderBoardRoute;
